Hazard) correction
Added a cell to edit the word Hazard) to Hazard. Added an early cell copy of the dataframe so that whenever we needed to rerun the whole code, it would run smoothly even after the edits. Otherwise it would mark an error when you edited something that was already edited to the original dataframe.